requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic-settings==2.6.1
httpx==0.28.0
pdf2image==1.17.0
Pillow==11.0.0
anthropic==0.40.0
redis==5.2.0
python-dotenv==1.0.1
